+++ src/osm_easy_api/api/_URLs.py
@@ -47,7 +47,8 @@ class URLs:
         }
 
         self.gpx: Dict[str, str] = {
-            "get": six_url + "/trackpoints?bbox={left},{bottom},{right},{top}&page={page_number}"
+            "get": six_url + "/trackpoints?bbox={left},{bottom},{right},{top}&page={page_number}",
+            "create": six_url + "/gpx/create"
         }
 
         self.user: Dict[str, str] = {